APRS-killer, APRS-logger: Added several stations

+++ LogKeeper/PreambleProcessing/APRS/APRS-killer.sh
@@ -6,10 +6,17 @@ LIB_BASE=${BASE}/lib
 export PYTHONPATH=${PYTHONPATH}:${LIB_BASE}/python-ogn-client
 
 file_base=$(date "+%Y-%m-%d")
-DATA_PATH=${BASE}/Data/YNRG/APRS
-
-for locker in $(ls ${DATA_PATH}/*.lock)
+DATA_BASE=${BASE}/Data
+for station in $(ls -d *)
 do
-  kill -SIGINT $(cat ${locker})
-  rm ${locker}
+  if [ -d ${station} ]
+  then
+    data_path=${DATA_BASE}/${station}/APRS
+    for locker in $(ls ${data_path}/*.lock 2> /dev/null)
+    do
+      kill -SIGINT $(cat ${locker})
+      kill -SIGHUP $(cat ${locker})
+      rm ${locker}
+    done
+  fi
 done
